Output the actual attachment name
I think the current code is designed with client-chosen default names in mind
but seems to output names that aren't actually used (likely because options[:name]
is passed in instead of the generated name). For now, just change messaging to
avoid saying anything about the name until we know what the name actually is.