Smarter add-on resolution
It can now resolve an add-on by a number of possible user-provided
identifiers:

* UUID
* Resource name (with or without leading @ symbol)
* Resource service + app name (if unambiguous)
* Resource plan + app name (if unambiguous)
* Attachment name + app name
* Attachment name *with* app name (e.g. acme-www::DATABASE)
  (this includes matching by substring, as is common with the pg:*
  commands)

The attachment resolution logic used for the the above is also broken
out into its own routine which can be useful in cases when targeting an
attachment itself (e.g. addons:detach or pg:promote, in the future).