Replace pg:promote command with one that uses the attachments API
It creates an attachment named DATABASE and forcefully replaces any
existing ones. If the attachment to be replaced is the only attachment
for its resource, the API will still reject the attachment request, so
there is no risk of orphaning a database or triggering a de-provision.

Optionally, we could only pass in the `confirm` parameter when user
actually provides it, but this would be a change in behaviour from the
existing command.